+++ src/caribou/__init__.py
@@ -1,0 +1,1 @@
+from .caribou_board import CaribouBoard  # noqa: F401