+++ src/caribou/current_bias.py
@@ -1,0 +1,5 @@
+from caribou.supply import Supply
+
+
+class CurrentBias(Supply):
+    """A current bias."""