+++ src/caribou/voltage_bias.py
@@ -1,0 +1,5 @@
+from caribou.supply import Supply
+
+
+class VoltageBias(Supply):
+    """A voltage bias."""